[Exporter] Refactor UC, SQL and SCIM objects into separate files
