+++ exporter/importables.go
@@ -5,7 +5,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"regexp"
 	"sort"
@@ -13,9 +12,7 @@ import (
 	"strings"
 
 	"github.com/databricks/databricks-sdk-go/apierr"
-	"github.com/databricks/databricks-sdk-go/service/catalog"
 	"github.com/databricks/databricks-sdk-go/service/compute"
-	"github.com/databricks/databricks-sdk-go/service/dashboards"
 	"github.com/databricks/databricks-sdk-go/service/iam"
 	"github.com/databricks/databricks-sdk-go/service/ml"
 	"github.com/databricks/databricks-sdk-go/service/pipelines"
@@ -25,23 +22,20 @@ import (
 	"github.com/databricks/databricks-sdk-go/service/sql"
 	"github.com/databricks/databricks-sdk-go/service/vectorsearch"
 	sdk_workspace "github.com/databricks/databricks-sdk-go/service/workspace"
-	tf_uc "github.com/databricks/terraform-provider-databricks/catalog"
 	"github.com/databricks/terraform-provider-databricks/common"
 	"github.com/databricks/terraform-provider-databricks/mws"
 	"github.com/databricks/terraform-provider-databricks/permissions/entity"
 	tf_dlt "github.com/databricks/terraform-provider-databricks/pipelines"
 	"github.com/databricks/terraform-provider-databricks/repos"
-	tfsettings "github.com/databricks/terraform-provider-databricks/settings"
+	tf_settings "github.com/databricks/terraform-provider-databricks/settings"
 	tf_sharing "github.com/databricks/terraform-provider-databricks/sharing"
 	tf_sql "github.com/databricks/terraform-provider-databricks/sql"
-	tf_sql_api "github.com/databricks/terraform-provider-databricks/sql/api"
 	"github.com/databricks/terraform-provider-databricks/storage"
 	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/zclconf/go-cty/cty"
 	"golang.org/x/exp/maps"
-	"golang.org/x/exp/slices"
 )
 
 var (
@@ -748,152 +742,10 @@ var resourcesMap map[string]importable = map[string]importable{
 		Service:        "groups",
 		WorkspaceLevel: true,
 		AccountLevel:   true,
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			return d.Get("display_name").(string) + "_" + d.Id()
-		},
-		List: func(ic *importContext) error {
-			if err := ic.cacheGroups(); err != nil {
-				return err
-			}
-			for offset, g := range ic.allGroups {
-				if !ic.MatchesName(g.DisplayName) {
-					log.Printf("[INFO] Group %s doesn't match %s filter", g.DisplayName, ic.match)
-					continue
-				}
-				ic.Emit(&resource{
-					Resource: "databricks_group",
-					ID:       g.ID,
-				})
-				log.Printf("[INFO] Scanned %d of %d groups", offset+1, len(ic.allGroups))
-			}
-			return nil
-		},
-		Search: func(ic *importContext, r *resource) error {
-			if r.Attribute != "display_name" {
-				return fmt.Errorf("wrong search attribute '%s' for databricks_group", r.Attribute)
-			}
-			if err := ic.cacheGroups(); err != nil {
-				return err
-			}
-			for _, g := range ic.allGroups {
-				if g.DisplayName == r.Value {
-					r.ID = g.ID
-					return nil
-				}
-			}
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			groupName := r.Data.Get("display_name").(string)
-			if (!ic.accountLevel && (groupName == "admins" || groupName == "users")) ||
-				(ic.accountLevel && groupName == "account users") {
-				// Workspace admins & users or Account users are to be imported through "data block"
-				r.Mode = "data"
-				r.Data.Set("workspace_access", false)
-				r.Data.Set("databricks_sql_access", false)
-				r.Data.Set("allow_instance_pool_create", false)
-				r.Data.Set("allow_cluster_create", false)
-				r.Data.State().Set(&terraform.InstanceState{
-					ID: r.ID,
-					Attributes: map[string]string{
-						"display_name": r.Name,
-					},
-				})
-			} else if r.Data != nil {
-				r.Data.Set("force", true)
-			}
-			if err := ic.cacheGroups(); err != nil {
-				return err
-			}
-			for _, g := range ic.allGroups {
-				if r.ID != g.ID {
-					continue
-				}
-				ic.emitRoles("group", g.ID, g.Roles)
-				builtInUserGroup := (ic.accountLevel && g.DisplayName == "account users") || (!ic.accountLevel && g.DisplayName == "users")
-				if builtInUserGroup && !ic.importAllUsers {
-					log.Printf("[INFO] Skipping import of entire user directory ...")
-					continue
-				}
-				if len(g.Members) > 10 {
-					log.Printf("[INFO] Importing %d members of %s", len(g.Members), g.DisplayName)
-				}
-				for _, parent := range g.Groups {
-					ic.Emit(&resource{
-						Resource: "databricks_group",
-						ID:       parent.Value,
-					})
-					if parent.Type == "direct" {
-						id := fmt.Sprintf("%s|%s", parent.Value, g.ID)
-						ic.Emit(&resource{
-							Resource: "databricks_group_member",
-							ID:       id,
-							Name:     fmt.Sprintf("%s_%s_%s", parent.Display, parent.Value, g.DisplayName),
-							Data:     ic.makeGroupMemberData(id, parent.Value, g.ID),
-						})
-					}
-				}
-				for i, x := range g.Members {
-					if strings.HasPrefix(x.Ref, "Users/") {
-						ic.Emit(&resource{
-							Resource: "databricks_user",
-							ID:       x.Value,
-						})
-						if !builtInUserGroup {
-							id := fmt.Sprintf("%s|%s", g.ID, x.Value)
-							ic.Emit(&resource{
-								Resource: "databricks_group_member",
-								ID:       id,
-								Name:     fmt.Sprintf("%s_%s_%s_%s", g.DisplayName, g.ID, x.Display, x.Value),
-								Data:     ic.makeGroupMemberData(id, g.ID, x.Value),
-							})
-						}
-					}
-					if strings.HasPrefix(x.Ref, "ServicePrincipals/") {
-						ic.Emit(&resource{
-							Resource: "databricks_service_principal",
-							ID:       x.Value,
-						})
-						if !builtInUserGroup {
-							id := fmt.Sprintf("%s|%s", g.ID, x.Value)
-							ic.Emit(&resource{
-								Resource: "databricks_group_member",
-								ID:       id,
-								Name:     fmt.Sprintf("%s_%s_%s_%s", g.DisplayName, g.ID, x.Display, x.Value),
-								Data:     ic.makeGroupMemberData(id, g.ID, x.Value),
-							})
-						}
-					}
-					if strings.HasPrefix(x.Ref, "Groups/") {
-						ic.Emit(&resource{
-							Resource: "databricks_group",
-							ID:       x.Value,
-						})
-						if !builtInUserGroup {
-							id := fmt.Sprintf("%s|%s", g.ID, x.Value)
-							ic.Emit(&resource{
-								Resource: "databricks_group_member",
-								ID:       id,
-								Name:     fmt.Sprintf("%s_%s_%s_%s", g.DisplayName, g.ID, x.Display, x.Value),
-								Data:     ic.makeGroupMemberData(id, g.ID, x.Value),
-							})
-						}
-					}
-					if len(g.Members) > 10 {
-						log.Printf("[INFO] Imported %d of %d members of %s", i+1, len(g.Members), g.DisplayName)
-					}
-				}
-			}
-			if ic.accountLevel {
-				ic.Emit(&resource{
-					Resource: "databricks_access_control_rule_set",
-					ID: fmt.Sprintf("accounts/%s/groups/%s/ruleSets/default",
-						ic.Client.Config.AccountID, r.ID),
-				})
-			}
-
-			return nil
-		},
+		Name:           makeNamePlusIdFunc("display_name"),
+		List:           listGroups,
+		Search:         searchGroup,
+		Import:         importGroup,
 	},
 	"databricks_group_member": {
 		Service:        "groups",
@@ -918,39 +770,9 @@ var resourcesMap map[string]importable = map[string]importable{
 			}
 			return nameNormalizationRegex.ReplaceAllString(strings.Split(s, "@")[0], "_") + "_" + d.Id()
 		},
-		List: func(ic *importContext) error {
-			ic.getUsersMapping()
-			ic.allUsersMutex.RLocker().Lock()
-			userMapping := maps.Clone(ic.allUsersMapping)
-			ic.allUsersMutex.RLocker().Unlock()
-			for userName, userScimId := range userMapping {
-				log.Printf("[TRACE] Emitting user %s, SCIM id=%s", userName, userScimId)
-				ic.Emit(&resource{
-					Resource: "databricks_user",
-					ID:       userScimId,
-				})
-			}
-			return nil
-		},
-		Search: func(ic *importContext, r *resource) error {
-			u, err := ic.findUserByName(r.Value, false)
-			if err != nil {
-				return err
-			}
-			r.ID = u.ID
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			username := r.Data.Get("user_name").(string)
-			r.Data.Set("force", true)
-			u, err := ic.findUserByName(username, false)
-			if err != nil {
-				return err
-			}
-			ic.emitGroups(*u)
-			ic.emitRoles("user", u.ID, u.Roles)
-			return nil
-		},
+		List:   listUsers,
+		Search: searchUser,
+		Import: importUser,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			if pathString == "display_name" {
 				userName := d.Get("user_name").(string)
@@ -974,28 +796,9 @@ var resourcesMap map[string]importable = map[string]importable{
 			}
 			return name + "_" + d.Id()
 		},
-		List: func(ic *importContext) error {
-			ic.getSpsMapping()
-			ic.spsMutex.RLock()
-			spsMapping := maps.Clone(ic.allSpsMapping)
-			ic.spsMutex.RLocker().Unlock()
-			for applicationId, appScimId := range spsMapping {
-				log.Printf("[TRACE] Emitting service principal %s, SCIM id=%s", applicationId, appScimId)
-				ic.Emit(&resource{
-					Resource: "databricks_service_principal",
-					ID:       appScimId,
-				})
-			}
-			return nil
-		},
-		Search: func(ic *importContext, r *resource) error {
-			u, err := ic.findSpnByAppID(r.Value, false)
-			if err != nil {
-				return err
-			}
-			r.ID = u.ID
-			return nil
-		},
+		List:   listServicePrincipals,
+		Search: searchServicePrincipal,
+		Import: importServicePrincipal,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			if pathString == "display_name" {
 				if ic.Client.IsAzure() {
@@ -1006,28 +809,11 @@ var resourcesMap map[string]importable = map[string]importable{
 				return false
 			}
 			// application_id should be provided only on Azure
+			// TODO: how to support Databricks-managed SPs?
 			if pathString == "application_id" {
 				return !ic.Client.IsAzure()
 			}
 			return defaultShouldOmitFieldFunc(ic, pathString, as, d)
-		},
-		Import: func(ic *importContext, r *resource) error {
-			applicationID := r.Data.Get("application_id").(string)
-			r.Data.Set("force", true)
-			u, err := ic.findSpnByAppID(applicationID, false)
-			if err != nil {
-				return err
-			}
-			ic.emitGroups(*u)
-			ic.emitRoles("service_principal", u.ID, u.Roles)
-			if ic.accountLevel {
-				ic.Emit(&resource{
-					Resource: "databricks_access_control_rule_set",
-					ID: fmt.Sprintf("accounts/%s/servicePrincipals/%s/ruleSets/default",
-						ic.Client.Config.AccountID, applicationID),
-				})
-			}
-			return nil
 		},
 	},
 	"databricks_permissions": {
@@ -1264,13 +1050,7 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_global_init_script": {
 		WorkspaceLevel: true,
 		Service:        "wsconf",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			name := d.Get("name").(string)
-			if name == "" {
-				return d.Id()
-			}
-			return name
-		},
+		Name:           makeNameOrIdFunc("name"),
 		List: func(ic *importContext) error {
 			globalInitScripts, err := ic.workspaceClient.GlobalInitScripts.ListAll(ic.Context)
 			if err != nil {
@@ -1581,66 +1361,9 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_query": {
 		WorkspaceLevel: true,
 		Service:        "queries",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			return d.Get("display_name").(string) + "_" + d.Id()
-		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.Queries.List(ic.Context, sql.ListQueriesRequest{PageSize: 100})
-			i := 0
-			for it.HasNext(ic.Context) {
-				q, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				i++
-				if !ic.MatchesName(q.DisplayName) {
-					continue
-				}
-				// TODO: look if we can create data based on the response, without calling Get
-				ic.EmitIfUpdatedAfterIsoString(&resource{
-					Resource:    "databricks_query",
-					ID:          q.Id,
-					Incremental: ic.incremental,
-				}, q.UpdateTime, fmt.Sprintf("query '%s'", q.DisplayName))
-				if i%50 == 0 {
-					log.Printf("[INFO] Imported %d Queries", i)
-				}
-			}
-			log.Printf("[INFO] Listed %d Queries", i)
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			var query tf_sql.QueryStruct
-			s := ic.Resources["databricks_query"].Schema
-			common.DataToStructPointer(r.Data, s, &query)
-			if query.WarehouseId != "" {
-				ic.Emit(&resource{
-					Resource: "databricks_sql_endpoint",
-					ID:       query.WarehouseId,
-				})
-			}
-			// emit queries specified as parameters
-			for _, p := range query.Parameters {
-				if p.QueryBackedValue != nil {
-					ic.Emit(&resource{
-						Resource: "databricks_query",
-						ID:       p.QueryBackedValue.QueryId,
-					})
-				}
-			}
-			ic.emitUserOrServicePrincipal(query.OwnerUserName)
-			ic.emitDirectoryOrRepo(query.ParentPath)
-			// TODO: r.AddExtraData(ParentDirectoryExtraKey, directoryPath) ?
-			ic.emitPermissionsIfNotIgnored(r, fmt.Sprintf("/sql/queries/%s", r.ID),
-				"query_"+ic.Importables["databricks_query"].Name(ic, r.Data))
-			if query.Catalog != "" && query.Schema != "" {
-				ic.Emit(&resource{
-					Resource: "databricks_schema",
-					ID:       fmt.Sprintf("%s.%s", query.Catalog, query.Schema),
-				})
-			}
-			return nil
-		},
+		Name:           makeNamePlusIdFunc("display_name"),
+		List:           listQueries,
+		Import:         importQuery,
 		// TODO: exclude owner if it's the current user?
 		Ignore: generateIgnoreObjectWithEmptyAttributeValue("databricks_query", "display_name"),
 		Depends: []reference{
@@ -1661,45 +1384,10 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_sql_endpoint": {
 		WorkspaceLevel: true,
 		Service:        "sql-endpoints",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			name := d.Get("name").(string)
-			if name == "" {
-				name = d.Id()
-			}
-			return name
-		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.Warehouses.List(ic.Context, sql.ListWarehousesRequest{})
-			i := 0
-			for it.HasNext(ic.Context) {
-				q, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				if !ic.MatchesName(q.Name) {
-					continue
-				}
-				ic.Emit(&resource{
-					Resource: "databricks_sql_endpoint",
-					ID:       q.Id,
-				})
-				i++
-				log.Printf("[INFO] Imported %d SQL endpoints", i)
-			}
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			ic.emitPermissionsIfNotIgnored(r, fmt.Sprintf("/sql/warehouses/%s", r.ID),
-				"sql_endpoint_"+ic.Importables["databricks_sql_endpoint"].Name(ic, r.Data))
-			if ic.meAdmin {
-				ic.Emit(&resource{
-					Resource: "databricks_sql_global_config",
-					ID:       tf_sql.GlobalSqlConfigResourceID,
-				})
-			}
-			return nil
-		},
-		Ignore: generateIgnoreObjectWithEmptyAttributeValue("databricks_sql_endpoint", "name"),
+		Name:           makeNameOrIdFunc("name"),
+		List:           listSqlEndpoints,
+		Import:         importSqlEndpoint,
+		Ignore:         generateIgnoreObjectWithEmptyAttributeValue("databricks_sql_endpoint", "name"),
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			switch pathString {
 			case "enable_serverless_compute":
@@ -1758,91 +1446,9 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_sql_dashboard": {
 		WorkspaceLevel: true,
 		Service:        "sql-dashboards",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			return d.Get("name").(string) + "_" + d.Id()
-		},
-		List: func(ic *importContext) error {
-			qs, err := dbsqlListObjects(ic, "/preview/sql/dashboards")
-			if err != nil {
-				return nil
-			}
-			for i, q := range qs {
-				name := q["name"].(string)
-				if !ic.MatchesName(name) {
-					continue
-				}
-				ic.EmitIfUpdatedAfterIsoString(&resource{
-					Resource:    "databricks_sql_dashboard",
-					ID:          q["id"].(string),
-					Incremental: ic.incremental,
-				}, q["updated_at"].(string), fmt.Sprintf("dashboard '%s'", name))
-				log.Printf("[INFO] Imported %d of %d SQL dashboards", i+1, len(qs))
-			}
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			ic.emitPermissionsIfNotIgnored(r, fmt.Sprintf("/sql/dashboards/%s", r.ID),
-				"sql_dashboard_"+ic.Importables["databricks_sql_dashboard"].Name(ic, r.Data))
-			dashboardID := r.ID
-			dashboardAPI := tf_sql.NewDashboardAPI(ic.Context, ic.Client)
-			dashboard, err := dashboardAPI.Read(dashboardID)
-			if err != nil {
-				return err
-			}
-
-			ic.emitSqlParentDirectory(dashboard.Parent)
-			for _, rv := range dashboard.Widgets {
-				var widget tf_sql_api.Widget
-				err = json.Unmarshal(rv, &widget)
-				if err != nil {
-					log.Printf("[WARN] Problems decoding widget for dashboard with ID: %s", dashboardID)
-					continue
-				}
-				widgetID := dashboardID + "/" + widget.ID.String()
-				ic.Emit(&resource{
-					Resource: "databricks_sql_widget",
-					ID:       widgetID,
-				})
-
-				if widget.VisualizationID != nil {
-					var visualization tf_sql_api.Visualization
-					err = json.Unmarshal(widget.Visualization, &visualization)
-					if err != nil {
-						log.Printf("[WARN] Problems decoding visualization for widget with ID: %s", widget.ID.String())
-						continue
-					}
-					if len(visualization.Query) > 0 {
-						var query tf_sql_api.Query
-						err = json.Unmarshal(visualization.Query, &query)
-						if err != nil {
-							log.Printf("[WARN] Problems decoding query for visualization with ID: %s", visualization.ID.String())
-							continue
-						}
-						visualizationID := query.ID + "/" + visualization.ID.String()
-						ic.Emit(&resource{
-							Resource: "databricks_sql_visualization",
-							ID:       visualizationID,
-						})
-						ic.Emit(&resource{
-							Resource: "databricks_query",
-							ID:       query.ID,
-						})
-						sqlEndpointID, err := ic.getSqlEndpoint(query.DataSourceID)
-						if err != nil {
-							log.Printf("[WARN] Can't find SQL endpoint for data source id %s", query.DataSourceID)
-						} else {
-							ic.Emit(&resource{
-								Resource: "databricks_sql_endpoint",
-								ID:       sqlEndpointID,
-							})
-						}
-					} else {
-						log.Printf("[DEBUG] Empty query in visualization %v", visualization)
-					}
-				}
-			}
-			return nil
-		},
+		Name:           makeNamePlusIdFunc("name"),
+		List:           listRedashDashboards,
+		Import:         importRedashDashboard,
 		Depends: []reference{
 			{Path: "parent", Resource: "databricks_directory", Match: "object_id", MatchType: MatchRegexp,
 				Regexp: sqlParentRegexp},
@@ -1862,10 +1468,7 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_sql_visualization": {
 		WorkspaceLevel: true,
 		Service:        "sql-dashboards",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			name := d.Get("name").(string) + "_" + d.Id()
-			return name
-		},
+		Name:           makeNamePlusIdFunc("name"),
 		Depends: []reference{
 			{Path: "query_id", Resource: "databricks_query"},
 		},
@@ -1873,34 +1476,8 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_alert": {
 		WorkspaceLevel: true,
 		Service:        "alerts",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			return d.Get("display_name").(string) + "_" + d.Id()
-		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.Alerts.List(ic.Context, sql.ListAlertsRequest{PageSize: 100})
-			i := 0
-			for it.HasNext(ic.Context) {
-				a, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				i++
-				if !ic.MatchesName(a.DisplayName) {
-					continue
-				}
-				// TODO: look if we can create data based on the response, without calling Get
-				ic.EmitIfUpdatedAfterIsoString(&resource{
-					Resource:    "databricks_alert",
-					ID:          a.Id,
-					Incremental: ic.incremental,
-				}, a.UpdateTime, fmt.Sprintf("alert '%s'", a.DisplayName))
-				if i%50 == 0 {
-					log.Printf("[INFO] Imported %d Alerts", i)
-				}
-			}
-			log.Printf("[INFO] Listed %d Alerts", i)
-			return nil
-		},
+		Name:           makeNamePlusIdFunc("display_name"),
+		List:           listAlerts,
 		Import: func(ic *importContext, r *resource) error {
 			var alert sql.Alert
 			s := ic.Resources["databricks_alert"].Schema
@@ -2430,62 +2007,12 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_system_schema": {
 		WorkspaceLevel: true,
 		Service:        "uc-system-schemas",
-		List: func(ic *importContext) error {
-			if ic.currentMetastore == nil {
-				return fmt.Errorf("there is no UC metastore information")
-			}
-			currentMetastore := ic.currentMetastore.MetastoreId
-			systemSchemas, err := ic.workspaceClient.SystemSchemas.ListAll(ic.Context,
-				catalog.ListSystemSchemasRequest{MetastoreId: currentMetastore})
-			if err != nil {
-				return err
-			}
-			for _, v := range systemSchemas {
-				if v.Schema == "information_schema" || v.Schema == "__internal_logging" {
-					continue
-				}
-				if v.State == catalog.SystemSchemaInfoStateEnableCompleted || v.State == catalog.SystemSchemaInfoStateEnableInitialized {
-					id := fmt.Sprintf("%s|%s", currentMetastore, v.Schema)
-					data := ic.Resources["databricks_system_schema"].Data(
-						&terraform.InstanceState{
-							ID: id,
-							Attributes: map[string]string{
-								"metastore_id": currentMetastore,
-								"schema":       v.Schema,
-							},
-						})
-					ic.Emit(&resource{
-						Resource: "databricks_system_schema",
-						ID:       id,
-						Data:     data,
-						Name:     nameNormalizationRegex.ReplaceAllString(id, "_"),
-					})
-				} else {
-					log.Printf("[DEBUG] Skipping system schema %s with state %s", v.Schema, v.State)
-				}
-			}
-			return nil
-		},
+		List:           listSystemSchemas,
 	},
 	"databricks_artifact_allowlist": {
 		WorkspaceLevel: true,
 		Service:        "uc-artifact-allowlist",
-		List: func(ic *importContext) error {
-			if ic.currentMetastore == nil {
-				return fmt.Errorf("there is no UC metastore information")
-			}
-			artifactTypes := []string{"INIT_SCRIPT", "LIBRARY_JAR", "LIBRARY_MAVEN"}
-			for _, v := range artifactTypes {
-				id := fmt.Sprintf("%s|%s", ic.currentMetastore.MetastoreId, v)
-				name := fmt.Sprintf("%s_%s_%s", v, ic.currentMetastore.Name, ic.currentMetastore.MetastoreId[:8])
-				ic.Emit(&resource{
-					Resource: "databricks_artifact_allowlist",
-					ID:       id,
-					Name:     nameNormalizationRegex.ReplaceAllString(name, "_"),
-				})
-			}
-			return nil
-		},
+		List:           listArtifactAllowLists,
 		Ignore: func(ic *importContext, r *resource) bool {
 			numBlocks := r.Data.Get("artifact_matcher.#").(int)
 			if numBlocks == 0 {
@@ -2508,75 +2035,8 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_catalog": {
 		WorkspaceLevel: true,
 		Service:        "uc-catalogs",
-		List: func(ic *importContext) error {
-			if ic.currentMetastore == nil {
-				return fmt.Errorf("there is no UC metastore information")
-			}
-			it := ic.workspaceClient.Catalogs.List(ic.Context, catalog.ListCatalogsRequest{})
-			for it.HasNext(ic.Context) {
-				v, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				switch v.CatalogType {
-				case "MANAGED_CATALOG", "FOREIGN_CATALOG", "DELTASHARING_CATALOG":
-					{
-						name := fmt.Sprintf("%s_%s_%s", v.Name, ic.currentMetastore.Name, v.CatalogType)
-						ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
-							Resource: "databricks_catalog",
-							ID:       v.Name,
-							Name:     nameNormalizationRegex.ReplaceAllString(name, "_"),
-						}, v.Name, v.UpdatedAt, fmt.Sprintf("catalog '%s'", v.Name))
-					}
-				default:
-					log.Printf("[INFO] Skipping catalog %s of type %s", v.Name, v.CatalogType)
-				}
-			}
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			var cat tf_uc.CatalogInfo
-			s := ic.Resources["databricks_catalog"].Schema
-			common.DataToStructPointer(r.Data, s, &cat)
-
-			// Emit: UC Connection, List schemas, Catalog grants, ...
-			owner, catalogGrantsResource := ic.emitUCGrantsWithOwner("catalog/"+cat.Name, r)
-			dependsOn := []*resource{}
-			if owner != "" && owner != ic.meUserName {
-				dependsOn = append(dependsOn, catalogGrantsResource)
-			}
-			// TODO: emit owner?  Should we do this? Because it's a account-level identity... Create a separate function for that...
-			if cat.ConnectionName != "" {
-				ic.Emit(&resource{
-					Resource: "databricks_connection",
-					ID:       cat.MetastoreID + "|" + cat.ConnectionName,
-				})
-			} else if cat.ShareName == "" {
-				// TODO: We need to be careful here if we add more catalog types... Really we need to have CatalogType in resource
-				if ic.isServiceInListing("uc-schemas") {
-					ignoredSchemas := []string{"information_schema"}
-					it := ic.workspaceClient.Schemas.List(ic.Context, catalog.ListSchemasRequest{CatalogName: r.ID})
-					for it.HasNext(ic.Context) {
-						schema, err := it.Next(ic.Context)
-						if err != nil {
-							return err
-						}
-						if schema.CatalogType != "MANAGED_CATALOG" || slices.Contains(ignoredSchemas, schema.Name) {
-							continue
-						}
-						ic.EmitIfUpdatedAfterMillis(&resource{
-							Resource:  "databricks_schema",
-							ID:        schema.FullName,
-							DependsOn: dependsOn,
-						}, schema.UpdatedAt, fmt.Sprintf("schema '%s'", schema.FullName))
-					}
-				}
-			}
-			if cat.IsolationMode == "ISOLATED" {
-				ic.emitWorkspaceBindings("catalog", cat.Name)
-			}
-			return nil
-		},
+		List:           listUcCatalogs,
+		Import:         importUcCatalog,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			if pathString == "isolation_mode" {
 				return d.Get(pathString).(string) != "ISOLATED"
@@ -2592,113 +2052,9 @@ var resourcesMap map[string]importable = map[string]importable{
 		//   This will require addition of the databricks_catalog data source
 	},
 	"databricks_schema": {
-		WorkspaceLevel: true,
-		Service:        "uc-schemas",
-		Import: func(ic *importContext, r *resource) error {
-			schemaFullName := r.ID
-			catalogName := r.Data.Get("catalog_name").(string)
-			schemaName := r.Data.Get("name").(string)
-			owner, schemaGrantResource := ic.emitUCGrantsWithOwner("schema/"+schemaFullName, r)
-			dependsOn := []*resource{}
-			if owner != "" && owner != ic.meUserName {
-				dependsOn = append(dependsOn, schemaGrantResource)
-			}
-			// TODO: think if we need to emit upstream dependencies in case if we're going bottom-up
-			ic.Emit(&resource{
-				Resource: "databricks_catalog",
-				ID:       catalogName,
-			})
-			// r.AddDependsOn(&resource{Resource: "databricks_grants", ID: "catalog/" + catalogName})
-
-			// TODO: somehow add depends on catalog's grant...
-			// TODO: emit owner? See comment in catalog resource
-			if ic.isServiceInListing("uc-models") {
-				it := ic.workspaceClient.RegisteredModels.List(ic.Context,
-					catalog.ListRegisteredModelsRequest{
-						CatalogName: catalogName,
-						SchemaName:  schemaName,
-					})
-				for it.HasNext(ic.Context) {
-					model, err := it.Next(ic.Context)
-					if err != nil {
-						return err // TODO: should we continue?
-					}
-					ic.EmitIfUpdatedAfterMillis(&resource{
-						Resource:  "databricks_registered_model",
-						ID:        model.FullName,
-						DependsOn: dependsOn,
-					}, model.UpdatedAt, fmt.Sprintf("registered model '%s'", model.FullName))
-				}
-			}
-			if ic.isServiceInListing("uc-volumes") {
-				// list volumes
-				it := ic.workspaceClient.Volumes.List(ic.Context,
-					catalog.ListVolumesRequest{
-						CatalogName: catalogName,
-						SchemaName:  schemaName,
-					})
-				for it.HasNext(ic.Context) {
-					volume, err := it.Next(ic.Context)
-					if err != nil {
-						return err // TODO: should we continue?
-					}
-					ic.EmitIfUpdatedAfterMillis(&resource{
-						Resource:  "databricks_volume",
-						ID:        volume.FullName,
-						DependsOn: dependsOn,
-					}, volume.UpdatedAt, fmt.Sprintf("volume '%s'", volume.FullName))
-				}
-			}
-			isTablesListingEnabled := ic.isServiceInListing("uc-tables")
-			isOnlineTablesListingEnabled := ic.isServiceInListing("uc-online-tables")
-			isVectorSearchListingEnabled := ic.isServiceInListing("vector-search")
-			if isTablesListingEnabled || isOnlineTablesListingEnabled || isVectorSearchListingEnabled {
-				it := ic.workspaceClient.Tables.List(ic.Context, catalog.ListTablesRequest{
-					CatalogName: catalogName,
-					SchemaName:  schemaName,
-				})
-				for it.HasNext(ic.Context) {
-					table, err := it.Next(ic.Context)
-					if err != nil {
-						return err // TODO: should we continue?
-					}
-					switch table.TableType {
-					case "MANAGED", "EXTERNAL", "VIEW":
-						if isTablesListingEnabled {
-							ic.EmitIfUpdatedAfterMillis(&resource{
-								Resource:  "databricks_sql_table",
-								ID:        table.FullName,
-								DependsOn: dependsOn,
-							}, table.UpdatedAt, fmt.Sprintf("table '%s'", table.FullName))
-						}
-					case "FOREIGN":
-						// TODO: it's better to use SecurableKind if it will be added to the Go SDK
-						switch table.DataSourceFormat {
-						case "VECTOR_INDEX_FORMAT":
-							if isVectorSearchListingEnabled {
-								ic.Emit(&resource{
-									Resource: "databricks_vector_search_index",
-									ID:       table.FullName,
-								})
-							}
-						case "MYSQL_FORMAT":
-							if isOnlineTablesListingEnabled {
-								ic.EmitIfUpdatedAfterMillis(&resource{
-									Resource:  "databricks_online_table",
-									ID:        table.FullName,
-									DependsOn: dependsOn,
-								}, table.UpdatedAt, fmt.Sprintf("table '%s'", table.FullName))
-							}
-						default:
-							log.Printf("[DEBUG] Skipping foreign table %s of format %s", table.FullName, table.DataSourceFormat)
-						}
-					default:
-						log.Printf("[DEBUG] Skipping table %s of type %s", table.FullName, table.TableType)
-					}
-				}
-			}
-			return nil
-		},
+		WorkspaceLevel:  true,
+		Service:         "uc-schemas",
+		Import:          importUcSchema,
 		ShouldOmitField: shouldOmitForUnityCatalog,
 		Ignore:          generateIgnoreObjectWithEmptyAttributeValue("databricks_schema", "name"),
 		Depends: []reference{
@@ -2709,19 +2065,7 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_volume": {
 		WorkspaceLevel: true,
 		Service:        "uc-volumes",
-		Import: func(ic *importContext, r *resource) error {
-			volumeFullName := r.ID
-			ic.emitUCGrantsWithOwner("volume/"+volumeFullName, r)
-
-			schemaFullName := r.Data.Get("catalog_name").(string) + "." + r.Data.Get("schema_name").(string)
-			ic.Emit(&resource{
-				Resource: "databricks_schema",
-				ID:       schemaFullName,
-			})
-			// r.AddDependsOn(&resource{Resource: "databricks_grants", ID: "schema/" + schemaFullName})
-			// TODO: emit owner? See comment in catalog resource
-			return nil
-		},
+		Import:         importUcVolume,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			if pathString == "storage_location" {
 				return d.Get("volume_type").(string) == "MANAGED"
@@ -2741,19 +2085,8 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_sql_table": {
 		WorkspaceLevel: true,
 		Service:        "uc-tables",
-		Import: func(ic *importContext, r *resource) error {
-			tableFullName := r.ID
-			ic.emitUCGrantsWithOwner("table/"+tableFullName, r)
-			schemaFullName := r.Data.Get("catalog_name").(string) + "." + r.Data.Get("schema_name").(string)
-			ic.Emit(&resource{
-				Resource: "databricks_schema",
-				ID:       schemaFullName,
-			})
-			// r.AddDependsOn(&resource{Resource: "databricks_grants", ID: "schema/" + schemaFullName})
-			// TODO: emit owner? See comment in catalog resource
-			return nil
-		},
-		Ignore: generateIgnoreObjectWithEmptyAttributeValue("databricks_sql_table", "name"),
+		Import:         importSqlTable,
+		Ignore:         generateIgnoreObjectWithEmptyAttributeValue("databricks_sql_table", "name"),
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			switch pathString {
 			case "storage_location":
@@ -2776,51 +2109,7 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_grants": {
 		WorkspaceLevel: true,
 		Service:        "uc-grants",
-		// TODO: Should we try to make name unique?
-		// TODO: do we need to emit principals? Maybe only on account level? See comment for the owner...
-		Import: func(ic *importContext, r *resource) error {
-			if ic.meUserName == "" {
-				return nil
-			}
-			// https://docs.databricks.com/en/data-governance/unity-catalog/manage-privileges/privileges.html#privilege-types-by-securable-object-in-unity-catalog
-			var newPrivileges []string
-			for k, v := range grantsPrivilegesToAdd {
-				if r.Data.Get(k).(string) != "" {
-					newPrivileges = append(newPrivileges, v...)
-					break
-				}
-			}
-			if len(newPrivileges) == 0 {
-				return nil
-			}
-
-			owner, found := r.GetExtraData("owner")
-			if !found || owner == "" || owner == ic.meUserName {
-				// We don't need to change permissions if owner isn't set, or it's the same user
-				return nil
-			}
-
-			var pList tf_uc.PermissionsList
-			s := ic.Resources["databricks_grants"].Schema
-			common.DataToStructPointer(r.Data, s, &pList)
-			foundExisting := false
-			for i, v := range pList.Assignments {
-				if v.Principal == ic.meUserName {
-					pList.Assignments[i].Privileges = append(pList.Assignments[i].Privileges, newPrivileges...)
-					slices.Sort(pList.Assignments[i].Privileges)
-					pList.Assignments[i].Privileges = slices.Compact(pList.Assignments[i].Privileges)
-					foundExisting = true
-					break
-				}
-			}
-			if !foundExisting {
-				pList.Assignments = append(pList.Assignments, tf_uc.PrivilegeAssignment{
-					Principal:  ic.meUserName,
-					Privileges: newPrivileges,
-				})
-			}
-			return common.StructToData(pList, s, r.Data)
-		},
+		Import:         importUcGrants,
 		Depends: []reference{
 			{Path: "catalog", Resource: "databricks_catalog"},
 			{Path: "schema", Resource: "databricks_schema"},
@@ -2840,80 +2129,20 @@ var resourcesMap map[string]importable = map[string]importable{
 		},
 	},
 	"databricks_storage_credential": {
-		WorkspaceLevel: true,
-		Service:        "uc-storage-credentials",
-		Import: func(ic *importContext, r *resource) error {
-			if r.ID == "__databricks_managed_storage_credential" {
-				// it's created by default and can't be imported
-				// TODO: add check for "securable_kind":"STORAGE_CREDENTIAL_DB_AWS_IAM" when we get it in the credential
-				r.Mode = "data"
-				data := tf_uc.ResourceStorageCredential().ToResource().TestResourceData()
-				obj := tf_uc.StorageCredentialInfo{Name: r.ID}
-				r.Data = ic.generateNewData(data, "databricks_storage_credential", r.ID, obj)
-			}
-			ic.emitUCGrantsWithOwner("storage_credential/"+r.ID, r)
-			if r.Data != nil {
-				isolationMode := r.Data.Get("isolation_mode").(string)
-				if isolationMode == "ISOLATION_MODE_ISOLATED" {
-					ic.emitWorkspaceBindings("storage_credential", r.ID)
-				}
-			}
-			return nil
-		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.StorageCredentials.List(ic.Context, catalog.ListStorageCredentialsRequest{})
-			for it.HasNext(ic.Context) {
-				v, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
-					Resource: "databricks_storage_credential",
-					ID:       v.Name,
-				}, v.Name, v.UpdatedAt, fmt.Sprintf("storage credential %s", v.Name))
-			}
-			return nil
-		},
+		WorkspaceLevel:  true,
+		Service:         "uc-storage-credentials",
+		Import:          importUcStorageCredential,
+		List:            listUcStorageCredentials,
 		ShouldOmitField: shouldOmitWithIsolationMode,
 		Depends: []reference{
 			{Path: "azure_service_principal.client_secret", Variable: true},
 		},
 	},
 	"databricks_credential": {
-		WorkspaceLevel: true,
-		Service:        "uc-credentials",
-		Import: func(ic *importContext, r *resource) error {
-			ic.emitUCGrantsWithOwner("credential/"+r.ID, r)
-			if r.Data != nil {
-				isolationMode := r.Data.Get("isolation_mode").(string)
-				if isolationMode == "ISOLATION_MODE_ISOLATED" {
-					purpose := r.Data.Get("purpose").(string)
-					if purpose == "SERVICE" {
-						ic.emitWorkspaceBindings("credential", r.ID)
-					} else if purpose == "STORAGE" {
-						ic.emitWorkspaceBindings("storage_credential", r.ID)
-					}
-				}
-			}
-			return nil
-		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.Credentials.ListCredentials(ic.Context, catalog.ListCredentialsRequest{})
-			for it.HasNext(ic.Context) {
-				v, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				if v.Purpose == catalog.CredentialPurposeStorage {
-					continue // we're handling storage credentials separately
-				}
-				ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
-					Resource: "databricks_credential",
-					ID:       v.Name,
-				}, v.Name, v.UpdatedAt, fmt.Sprintf("credential %s", v.Name))
-			}
-			return nil
-		},
+		WorkspaceLevel:  true,
+		Service:         "uc-credentials",
+		Import:          importUcCredential,
+		List:            listUcCredentials,
 		ShouldOmitField: shouldOmitWithIsolationMode,
 		Depends: []reference{
 			{Path: "azure_service_principal.client_secret", Variable: true},
@@ -2922,46 +2151,8 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_external_location": {
 		WorkspaceLevel: true,
 		Service:        "uc-external-locations",
-		Import: func(ic *importContext, r *resource) error {
-			if r.ID == dbManagedExternalLocation {
-				// it's created by default and can't be imported
-				// TODO: add check for "securable_kind":"EXTERNAL_LOCATION_DB_STORAGE" when we get it in the credential
-				r.Mode = "data"
-				data := tf_uc.ResourceExternalLocation().ToResource().TestResourceData()
-				obj := tf_uc.ExternalLocationInfo{ExternalLocationInfo: catalog.ExternalLocationInfo{Name: r.ID}}
-				r.Data = ic.generateNewData(data, "databricks_external_location", r.ID, obj)
-			}
-			ic.emitUCGrantsWithOwner("external_location/"+r.ID, r)
-			credentialName := r.Data.Get("credential_name").(string)
-			ic.Emit(&resource{
-				Resource: "databricks_storage_credential",
-				ID:       credentialName,
-			})
-			if r.Data != nil {
-				isolationMode := r.Data.Get("isolation_mode").(string)
-				if isolationMode == "ISOLATION_MODE_ISOLATED" {
-					ic.emitWorkspaceBindings("external_location", r.ID)
-				}
-			}
-			// r.AddDependsOn(&resource{Resource: "databricks_grants", ID: "storage_credential/" + credentialName})
-			return nil
-		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.ExternalLocations.List(ic.Context, catalog.ListExternalLocationsRequest{})
-			for it.HasNext(ic.Context) {
-				v, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				if v.Name != "metastore_default_location" {
-					ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
-						Resource: "databricks_external_location",
-						ID:       v.Name,
-					}, v.Name, v.UpdatedAt, fmt.Sprintf("external location %s", v.Name))
-				}
-			}
-			return nil
-		},
+		Import:         importUcExternalLocation,
+		List:           listUcExternalLocations,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			if (pathString == "url" || pathString == "credential_name") && d.Get("name").(string) == dbManagedExternalLocation {
 				return true
@@ -2990,20 +2181,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			}
 			return connectionType + "_" + connectionName
 		},
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.Connections.List(ic.Context, catalog.ListConnectionsRequest{})
-			for it.HasNext(ic.Context) {
-				conn, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
-					Resource: "databricks_connection",
-					ID:       conn.MetastoreId + "|" + conn.Name,
-				}, conn.Name, conn.UpdatedAt, fmt.Sprintf("connection '%s'", conn.Name))
-			}
-			return nil
-		},
+		List: listUcConnections,
 		// TODO: think what to do with the sensitive fields in the `options`?
 		Import: func(ic *importContext, r *resource) error {
 			// TODO: do we need to emit the owner See comment for the owner...
@@ -3143,48 +2321,9 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_metastore": {
 		AccountLevel: true,
 		Service:      "uc-metastores",
-		Name: func(ic *importContext, d *schema.ResourceData) string {
-			name := d.Get("name").(string)
-			if name == "" {
-				return d.Id()
-			}
-			return name
-		},
-		List: func(ic *importContext) error {
-			it := ic.accountClient.Metastores.List(ic.Context)
-			for it.HasNext(ic.Context) {
-				mstore, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				ic.EmitIfUpdatedAfterMillisAndNameMatches(&resource{
-					Resource: "databricks_metastore",
-					ID:       mstore.MetastoreId,
-				}, mstore.Name, mstore.UpdatedAt, fmt.Sprintf("metastore '%s'", mstore.Name))
-			}
-			return nil
-		},
-		Import: func(ic *importContext, r *resource) error {
-			ic.emitUCGrantsWithOwner("metastore/"+r.ID, r)
-			// TODO: emit owner? See comment in catalog resource
-			if ic.accountLevel {
-				// emit metastore assignments
-				assignments, err := ic.accountClient.MetastoreAssignments.ListByMetastoreId(ic.Context, r.ID)
-				if err == nil {
-					for _, workspaceID := range assignments.WorkspaceIds {
-						ic.Emit(&resource{
-							Resource: "databricks_metastore_assignment",
-							ID:       fmt.Sprintf("%d|%s", workspaceID, r.ID),
-						})
-					}
-				} else {
-					log.Printf("[ERROR] listing metastore assignments: %s", err.Error())
-				}
-				// TODO: emit storage credentials associated with specific metastores, but we'll need to solve
-				// a problem of importing a resource... This will require to changing ID from name to metastore ID + name.
-			}
-			return nil
-		},
+		Name:         makeNameOrIdFunc("name"),
+		List:         listUcMetastores,
+		Import:       importUcMetastores,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			if pathString == "default_data_access_config_id" || pathString == "storage_root_credential_id" {
 				// technically, both should be marked as `computed`
@@ -3236,40 +2375,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			fileNameMd5 := fmt.Sprintf("%x", md5.Sum([]byte(name)))
 			return strings.ToLower(name) + "_" + fileNameMd5[:8]
 		},
-		Import: func(ic *importContext, r *resource) error {
-			parts := strings.Split(r.ID, "/")
-			// Converting /Volumes/<catalog>/<schema>/<table>/<file> to <catalog>.<schema>.<table>
-			if len(parts) > 5 {
-				volumeId := strings.Join(parts[2:5], ".")
-				ic.Emit(&resource{
-					Resource: "databricks_volume",
-					ID:       volumeId,
-				})
-				// r.AddDependsOn(&resource{Resource: "databricks_grants", ID: "volume/" + volumeId})
-			}
-
-			// download & store file
-			resp, err := ic.workspaceClient.Files.DownloadByFilePath(ic.Context, r.ID)
-			if err != nil {
-				return err
-			}
-			// write file
-			fileName := ic.prefix + fileNameNormalizationRegex.ReplaceAllString(strings.TrimPrefix(r.ID, "/Volumes/"), "_")
-			local, relativeName, err := ic.createFileIn("uc_files", fileName)
-			if err != nil {
-				return err
-			}
-			defer local.Close()
-			defer resp.Contents.Close()
-			_, err = io.Copy(local, resp.Contents)
-			if err != nil {
-				return err
-			}
-			r.Data.Set("source", relativeName)
-			r.Data.Set("path", r.ID)
-
-			return nil
-		},
+		Import: importUcVolumeFile,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			switch pathString {
 			case "md5", "remote_file_modified", "modification_time", "file_size":
@@ -3350,30 +2456,7 @@ var resourcesMap map[string]importable = map[string]importable{
 	"databricks_dashboard": {
 		WorkspaceLevel: true,
 		Service:        "dashboards",
-		List: func(ic *importContext) error {
-			it := ic.workspaceClient.Lakeview.List(ic.Context, dashboards.ListDashboardsRequest{PageSize: 1000})
-			i := 0
-			for it.HasNext(ic.Context) {
-				d, err := it.Next(ic.Context)
-				if err != nil {
-					return err
-				}
-				i++
-				if !ic.MatchesName(d.DisplayName) {
-					continue
-				}
-				// TODO: add emit for incremental mode. But this information isn't included into the List response
-				ic.Emit(&resource{
-					Resource: "databricks_dashboard",
-					ID:       d.DashboardId,
-				})
-				if i%100 == 0 {
-					log.Printf("[INFO] Processed %d dashboards", i)
-				}
-			}
-			log.Printf("[INFO] Listed %d dashboards", i)
-			return nil
-		},
+		List:           listLakeviewDashboards,
 		Name: func(ic *importContext, d *schema.ResourceData) string {
 			s := d.Get("parent_path").(string)
 			if s != "" {
@@ -3389,43 +2472,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			s = s + "_" + d.Id()
 			return nameNormalizationRegex.ReplaceAllString(s, "_")
 		},
-		Import: func(ic *importContext, r *resource) error {
-			path := r.Data.Get("path").(string)
-			if ic.isInRepoOrGitFolder(path, false) {
-				ic.emitRepoOrGitFolder(path, false)
-				return nil
-			}
-			parts := strings.Split(path, "/")
-			plen := len(parts)
-			if idx := strings.Index(parts[plen-1], "."); idx != -1 {
-				parts[plen-1] = parts[plen-1][:idx] + "_" + r.ID + parts[plen-1][idx:]
-			} else {
-				parts[plen-1] = parts[plen-1] + "_" + r.ID
-			}
-			name := fileNameNormalizationRegex.ReplaceAllString(strings.Join(parts, "/")[1:], "_")
-			fileName, err := ic.saveFileIn("dashboards", name, []byte(r.Data.Get("serialized_dashboard").(string)))
-			if err != nil {
-				return err
-			}
-			r.Data.Set("file_path", fileName)
-			r.Data.Set("serialized_dashboard", "")
-
-			ic.emitPermissionsIfNotIgnored(r, "/dashboards/"+r.ID,
-				"dashboard_"+ic.Importables["databricks_dashboard"].Name(ic, r.Data))
-			parentPath := r.Data.Get("parent_path").(string)
-			if parentPath != "" && parentPath != "/" {
-				ic.emitDirectoryOrRepo(parentPath)
-			}
-			warehouseId := r.Data.Get("warehouse_id").(string)
-			if warehouseId != "" {
-				ic.Emit(&resource{
-					Resource: "databricks_sql_endpoint",
-					ID:       warehouseId,
-				})
-			}
-
-			return nil
-		},
+		Import: importLakeviewDashboard,
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
 			return pathString == "dashboard_change_detected" || shouldOmitMd5Field(ic, pathString, as, d)
 		},
@@ -3472,7 +2519,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			return nil
 		},
 		Import: func(ic *importContext, r *resource) error {
-			var notificationDestination tfsettings.NDStruct
+			var notificationDestination tf_settings.NDStruct
 			s := ic.Resources["databricks_notification_destination"].Schema
 			common.DataToStructPointer(r.Data, s, &notificationDestination)
 			if notificationDestination.DestinationType == "EMAIL" && notificationDestination.Config != nil &&
@@ -3484,7 +2531,7 @@ var resourcesMap map[string]importable = map[string]importable{
 			return nil
 		},
 		ShouldOmitField: func(ic *importContext, pathString string, as *schema.Schema, d *schema.ResourceData) bool {
-			var notificationDestination tfsettings.NDStruct
+			var notificationDestination tf_settings.NDStruct
 			s := ic.Resources["databricks_notification_destination"].Schema
 			common.DataToStructPointer(d, s, &notificationDestination)
 			if notificationDestination.Config != nil {
